move domain classes up a package and fix imports
