+++ packages/syft/src/syft/core/node/domain.py
@@ -19,66 +19,66 @@ from nacl.signing import VerifyKey
 from pydantic import BaseSettings
 
 # relative
-from ....lib.python import String
-from ....logger import critical
-from ....logger import debug
-from ....logger import info
-from ....logger import traceback
-from ...adp.adversarial_accountant import AdversarialAccountant
-from ...common.message import SignedImmediateSyftMessageWithReply
-from ...common.message import SignedMessage
-from ...common.message import SyftMessage
-from ...common.uid import UID
-from ...io.location import Location
-from ...io.location import SpecificLocation
-from ..common.action.get_object_action import GetObjectAction
-from ..common.client import Client
-from ..common.node import Node
-from ..common.node_manager.association_request_manager import AssociationRequestManager
-from ..common.node_manager.dataset_manager import DatasetManager
-from ..common.node_manager.environment_manager import EnvironmentManager
-from ..common.node_manager.node_manager import NodeManager
-from ..common.node_manager.node_route_manager import NodeRouteManager
-from ..common.node_manager.redis_store import RedisStore
-from ..common.node_manager.request_manager import RequestManager
-from ..common.node_manager.role_manager import RoleManager
-from ..common.node_manager.user_manager import UserManager
-from ..common.node_service.association_request.association_request_service import (
+from ...lib.python import String
+from ...logger import critical
+from ...logger import debug
+from ...logger import info
+from ...logger import traceback
+from ..adp.adversarial_accountant import AdversarialAccountant
+from ..common.message import SignedImmediateSyftMessageWithReply
+from ..common.message import SignedMessage
+from ..common.message import SyftMessage
+from ..common.uid import UID
+from ..io.location import Location
+from ..io.location import SpecificLocation
+from .common.action.get_object_action import GetObjectAction
+from .common.client import Client
+from .common.node import Node
+from .common.node_manager.association_request_manager import AssociationRequestManager
+from .common.node_manager.dataset_manager import DatasetManager
+from .common.node_manager.environment_manager import EnvironmentManager
+from .common.node_manager.node_manager import NodeManager
+from .common.node_manager.node_route_manager import NodeRouteManager
+from .common.node_manager.redis_store import RedisStore
+from .common.node_manager.request_manager import RequestManager
+from .common.node_manager.role_manager import RoleManager
+from .common.node_manager.user_manager import UserManager
+from .common.node_service.association_request.association_request_service import (
     AssociationRequestService,
 )
-from ..common.node_service.dataset_manager.dataset_manager_service import (
+from .common.node_service.dataset_manager.dataset_manager_service import (
     DatasetManagerService,
 )
-from ..common.node_service.get_remaining_budget.get_remaining_budget_service import (
+from .common.node_service.get_remaining_budget.get_remaining_budget_service import (
     GetRemainingBudgetService,
 )
-from ..common.node_service.node_setup.node_setup_messages import (
+from .common.node_service.node_setup.node_setup_messages import (
     CreateInitialSetUpMessage,
 )
-from ..common.node_service.node_setup.node_setup_service import NodeSetupService
-from ..common.node_service.object_request.object_request_service import (
+from .common.node_service.node_setup.node_setup_service import NodeSetupService
+from .common.node_service.object_request.object_request_service import (
     ObjectRequestServiceWithoutReply,
 )
-from ..common.node_service.object_request.object_request_service import RequestService
-from ..common.node_service.ping.ping_service import PingService
-from ..common.node_service.publish.publish_service import PublishScalarsService
-from ..common.node_service.request_answer.request_answer_messages import RequestStatus
-from ..common.node_service.request_answer.request_answer_service import (
+from .common.node_service.object_request.object_request_service import RequestService
+from .common.node_service.ping.ping_service import PingService
+from .common.node_service.publish.publish_service import PublishScalarsService
+from .common.node_service.request_answer.request_answer_messages import RequestStatus
+from .common.node_service.request_answer.request_answer_service import (
     RequestAnswerService,
 )
-from ..common.node_service.request_receiver.request_receiver_messages import (
+from .common.node_service.request_receiver.request_receiver_messages import (
     RequestMessage,
 )
-from ..common.node_service.role_manager.role_manager_service import RoleManagerService
-from ..common.node_service.simple.simple_service import SimpleService
-from ..common.node_service.user_auth.user_auth_service import UserLoginService
-from ..common.node_service.user_manager.user_manager_service import UserManagerService
-from ..common.node_service.vpn.vpn_service import VPNConnectService
-from ..common.node_service.vpn.vpn_service import VPNJoinService
-from ..common.node_service.vpn.vpn_service import VPNStatusService
-from ..common.node_table.utils import create_memory_db_engine
-from ..device import Device
-from ..device import DeviceClient
+from .common.node_service.role_manager.role_manager_service import RoleManagerService
+from .common.node_service.simple.simple_service import SimpleService
+from .common.node_service.user_auth.user_auth_service import UserLoginService
+from .common.node_service.user_manager.user_manager_service import UserManagerService
+from .common.node_service.vpn.vpn_service import VPNConnectService
+from .common.node_service.vpn.vpn_service import VPNJoinService
+from .common.node_service.vpn.vpn_service import VPNStatusService
+from .common.node_table.utils import create_memory_db_engine
+from .device import Device
+from .device import DeviceClient
 from .domain_client import DomainClient
 from .domain_service import DomainServiceClass
