--- packages/syft/src/syft/core/node/domain/__init__.py
@@ -1,5 +1,0 @@
-# relative
-from .domain import Domain  # noqa: F401
-from .domain_client import DomainClient  # noqa: F401
-
-__all__ = ["DomainClient", "Domain"]
