+++ packages/syft/src/syft/core/node/enums.py
